# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Android Laravel",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/login/{nroDoc}/{pass}": {
            "get": {
                "description": "Iniciar sesión",
                "operationId": "inicioSesion",
                "parameters": [
                    {
                        "name": "nroDoc",
                        "in": "path",
                        "description": "Numero de documento",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "pass",
                        "in": "path",
                        "description": "Contraseña inicio de sesion",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Token del usuario."
                    }
                }
            }
        },
        "/crearUsuario": {
            "post": {
                "description": "Iniciar sesión",
                "operationId": "crearUsuario",
                "parameters": [
                    {
                        "name": "nroDoc",
                        "in": "path",
                        "description": "Numero de documento",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "String"
                        }
                    },
                    {
                        "name": "nombres",
                        "in": "path",
                        "description": "Nombres del usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellidos",
                        "in": "path",
                        "description": "Apellidos del usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "telefonos",
                        "in": "path",
                        "description": "Telefono del usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idCreador",
                        "in": "path",
                        "description": "Creador",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Se ha creado el usuario."
                    }
                }
            }
        },
        "/actualizarDatos": {
            "put": {
                "description": "Iniciar sesión",
                "operationId": "actualizarDatos",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "Nombres del usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellidos",
                        "in": "path",
                        "description": "Apellidos del usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "telefono",
                        "in": "path",
                        "description": "Telefono del usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id Usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Usuario actualizado correctamente."
                    }
                }
            }
        },
        "listaUsuarios/{tiempoToken}": {
            "get": {
                "description": "Iniciar sesión",
                "operationId": "actualizarDatos",
                "parameters": [
                    {
                        "name": "tiempoToken",
                        "in": "path",
                        "description": "Tiempo del token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Datos usuarios."
                    }
                }
            }
        },
        "/deshabilitarUsuario": {
            "put": {
                "description": "Deshabilitar usuarios",
                "operationId": "deshabilitarUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "200": {
                        "description": "Usuario deshabilitado."
                    }
                }
            }
        }
    }
}